refactor(redis): improve health monitoring and logging logic
- Initialized Redis health status to false to ensure the first success is logged.
- Enhanced logging to provide clearer messages when transitioning between healthy and unhealthy states.
- Consolidated the import of ClientType from a new socket-types module for better organization.
- Updated multiple protocol files to reflect the new import path for ClientType.
