+++ packages/sdk-socket-server-next/src/protocol/retrieveMessages.ts
@@ -1,6 +1,6 @@
 import { pubClient } from '../redis';
 import { getLogger } from '../logger';
-import { ClientType } from '../socket-config';
+import { ClientType } from '../socket-types';
 import { incrementKeyMigration } from '../metrics';
 import { QueuedMessage } from './handleMessage';
