+++ packages/sdk-socket-server-next/src/socket-types.ts
@@ -1,0 +1,1 @@
+export type ClientType = 'dapp' | 'wallet';
